requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiooss2==0.2.11
aiosignal==1.3.2
airportsdata==20250523
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-core-v3==2.13.33
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bitsandbytes==0.46.0
black==25.1.0
blake3==1.0.5
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
compressed-tensors==0.9.3
crcmod==1.7
cryptography==45.0.4
cupy-cuda12x==13.4.1
DataProperty==1.1.0
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
deepspeed==0.17.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
flake8==7.2.0
flash_attn==2.8.0.post2
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.2
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.0.0
iniconfig==2.1.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
lark==1.2.2
latex2sympy2_extended==1.0.6
liger_kernel==0.5.2
lighteval==0.10.0
llguidance==0.7.29
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.5.2
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mccabe==0.7.0
mdurl==0.1.2
mistral_common==1.6.2
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nebula-metadata-sdk==1.0.5
nebulactl==1.1.13
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.1
odps==3.5.1
-e git+http://gitlab.alibaba-inc.com/rs_ml/md_link_assessment.git@a4c43ae008110bc0193fa20f4028e02c5f0ba7a2#egg=open_r1&subdirectory=Visual-RFT/src/virft
openai==1.86.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openlm_hub==0.0.35
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
oss2==2.9.1
ossfs==2025.5.0
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.0
pangudfs-client==2.4.2+lightsdk1.1.28
parameterized==0.9.0
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pathspec==0.12.1
pathvalidate==3.3.1
pexpect==4.9.0
pfzy==0.3.4
pillow==11.2.1
pip-chill==1.0.3
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyodps==0.12.3
pytablewriter==1.2.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.11
ray==2.47.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.3.3
rich==14.0.0
rich-toolkit==0.14.7
rouge_score==0.1.2
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.30.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
style==1.1.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==2.3.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
trl==0.18.2
typepy==1.3.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update==0.0.1
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
vllm==0.8.4
wandb==0.18.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
